+++ NEWS.md
@@ -6,9 +6,9 @@ Phan NEWS
 Backwards incompatible changes:
 + Switch from AST version 70 to AST version 80.
   `php-ast` should be upgraded to version 1.0.10-dev or newer.
-+ Support analyzing PHP 8.0 attributes.
++ Support analyzing PHP 8.0 attributes when Phan is run with php 8.0 or newer.
   New issue types: `PhanCompatibleAttributeOnSameLine`, `PhanAttributeNonAttribute`, `PhanAttributeNonClass`, `PhanAttributeNonRepeatable`,
-  `PhanUndeclaredClassAttribute`.
+  `PhanUndeclaredClassAttribute`, `PhanAttributeWrongTarget`.
 + Drop the no-op `--polyfill-parse-all-doc-comments` flag.
 
 Miscellaneous: